+++ cralwer_plugin/base_plugin/src/main/java/com/trucy/novel/crawler/dto/CrawlerChapterDto.java
@@ -1,0 +1,16 @@
+package com.trucy.novel.crawler.dto;
+
+import lombok.*;
+
+@Data
+@Builder
+@ToString
+@EqualsAndHashCode
+@AllArgsConstructor
+@NoArgsConstructor
+public class CrawlerChapterDto {
+    private String id;
+    private String url;
+    private String name;
+    private String htmlContent;
+}
