+++ java/src/main/java/com/trucy/novel/RunApplication.java
@@ -1,0 +1,23 @@
+package com.trucy.novel;
+
+import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@Slf4j
+@EnableAsync
+@EnableScheduling
+@AllArgsConstructor
+@SpringBootApplication
+@EnableConfigurationProperties
+public class RunApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(RunApplication.class, args);
+    }
+
+}
