+++ java/src/main/java/com/trucy/novel/config/BusinessConfig.java
@@ -1,0 +1,12 @@
+package com.trucy.novel.config;
+
+import lombok.Data;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+
+@Data
+@Component
+@ConfigurationProperties(prefix = "config.business")
+public class BusinessConfig {
+}
