+++ java/src/main/java/com/trucy/novel/config/request/ApiResponse.java
@@ -1,0 +1,16 @@
+package com.trucy.novel.config.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class ApiResponse {
+    private int code;
+    private String message;
+    private Object data;
+}
