+++ java/src/main/java/com/trucy/novel/config/request/ex/BusinessException.java
@@ -1,0 +1,15 @@
+package com.trucy.novel.config.request.ex;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@EqualsAndHashCode(callSuper = true)
+public class BusinessException extends Exception {
+    private int errorCode;
+    private String errorMessage;
+}
