+++ java/src/main/java/com/trucy/novel/config/request/ex/CrawlerException.java
@@ -1,0 +1,14 @@
+package com.trucy.novel.config.request.ex;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@EqualsAndHashCode(callSuper = true)
+public class CrawlerException extends Exception {
+    private String errorMessage;
+}
