--- java/src/main/java/com/trucy/novel/crawler/config/BusinessConfig.java
@@ -1,12 +1,0 @@
-package com.trucy.novel.crawler.config;
-
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-
-@Data
-@Component
-@ConfigurationProperties(prefix = "config.business")
-public class BusinessConfig {
-}
