--- java/src/main/java/com/trucy/novel/crawler/config/GlobalConfiguration.java
@@ -1,17 +1,0 @@
-package com.trucy.novel.crawler.config;
-
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
-
-@Configuration
-public class GlobalConfiguration {
-
-    @Bean
-    public LocalValidatorFactoryBean getValidatorFactory() {
-        LocalValidatorFactoryBean localValidatorFactoryBean = new LocalValidatorFactoryBean();
-        localValidatorFactoryBean.getValidationPropertyMap().put("hibernate.validator.fail_fast", "true");
-        return localValidatorFactoryBean;
-    }
-}
