--- java/src/main/java/com/trucy/novel/crawler/config/request/ApiResponse.java
@@ -1,16 +1,0 @@
-package com.trucy.novel.crawler.config.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class ApiResponse {
-    private int code;
-    private String message;
-    private Object data;
-}
