--- java/src/main/java/com/trucy/novel/crawler/crawler/dto/CrawlerChapterDto.java
@@ -1,16 +1,0 @@
-package com.trucy.novel.crawler.crawler.dto;
-
-import lombok.*;
-
-@Data
-@Builder
-@ToString
-@EqualsAndHashCode
-@AllArgsConstructor
-@NoArgsConstructor
-public class CrawlerChapterDto {
-    private String id;
-    private String url;
-    private String name;
-    private String htmlContent;
-}
