+++ java/src/main/java/com/trucy/novel/utils/NanoIdUtils.java
@@ -1,0 +1,63 @@
+package com.trucy.novel.utils;
+
+import java.security.SecureRandom;
+import java.util.Random;
+
+public final class NanoIdUtils {
+    public static final SecureRandom DEFAULT_NUMBER_GENERATOR = new SecureRandom();
+    // 随机生成的字符内容
+    public static final char[] DEFAULT_ALPHABET = "0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ".toCharArray();
+    // 随机生成 ID 的位数
+    public static final int DEFAULT_SIZE = 15;
+
+    public static final char[] DEFAULT_NUMBER_ALPHABET = "1234567890".toCharArray();
+
+
+    private NanoIdUtils() {
+    }
+
+    public static String randomNanoId() {
+        return randomNanoId(DEFAULT_NUMBER_GENERATOR, DEFAULT_ALPHABET, DEFAULT_SIZE);
+    }
+
+    public static String randomNanoId(int size) {
+        return randomNanoId(DEFAULT_NUMBER_GENERATOR, DEFAULT_ALPHABET, size);
+    }
+
+    public static String randomNumberNanoId(int size) {
+        return randomNanoId(DEFAULT_NUMBER_GENERATOR, DEFAULT_NUMBER_ALPHABET, size);
+    }
+
+    public static String randomNanoId(Random random, char[] alphabet, int size) {
+        if (random == null) {
+            throw new IllegalArgumentException("random cannot be null.");
+        } else if (alphabet == null) {
+            throw new IllegalArgumentException("alphabet cannot be null.");
+        } else if (alphabet.length != 0 && alphabet.length < 256) {
+            if (size <= 0) {
+                throw new IllegalArgumentException("size must be greater than zero.");
+            } else {
+                int mask = (2 << (int) Math.floor(Math.log((double) (alphabet.length - 1)) / Math.log(2.0D))) - 1;
+                int step = (int) Math.ceil(1.6D * (double) mask * (double) size / (double) alphabet.length);
+                StringBuilder idBuilder = new StringBuilder();
+
+                while (true) {
+                    byte[] bytes = new byte[step];
+                    random.nextBytes(bytes);
+
+                    for (int i = 0; i < step; ++i) {
+                        int alphabetIndex = bytes[i] & mask;
+                        if (alphabetIndex < alphabet.length) {
+                            idBuilder.append(alphabet[alphabetIndex]);
+                            if (idBuilder.length() == size) {
+                                return idBuilder.toString();
+                            }
+                        }
+                    }
+                }
+            }
+        } else {
+            throw new IllegalArgumentException("alphabet must contain between 1 and 255 symbols.");
+        }
+    }
+}
